RSDK-110 RSDK-117 RSDK-144 Return object refactor
RSDK-110 Standardized the return objects of the basic API functions,
as well as many of the SDK package functions, to meet a standard
list(entity, response) structure (where appropriate). This change is
a major breaking change to the package, as most SDK functions needed
modification to handle the altered return object from apiPOST and apiPUT
and/or had modifications of their own return objects.
RSDK-117 Added "fullReturn" parameter to API functions to provide
option to not receive full HTTP response object in return.
RSDK-144 Separated set of API tests to individual files.